footer button and logo sizing
